+++ src/Smartstore.Core/Platform/DataExchange/Export/DataExporter.DynamicProduct.cs
@@ -99,7 +99,14 @@ namespace Smartstore.Core.DataExchange.Export
                 return null;
             }
 
-            dynamic result = new DynamicEntity(product);
+            var dynamicEntity = new DynamicEntity(product);
+
+            if (product.MergedDataValues != null)
+            {
+                dynamicEntity.MergeRange(product.MergedDataValues);
+            }
+
+            dynamic result = dynamicEntity;
             var localizedName = ctx.GetTranslation(product, nameof(product.Name), product.Name);
 
             result.AppliedDiscounts = null;