+++ changelog.md
@@ -47,6 +47,7 @@
 - Routing: non-slug, unlocalized system routes did not redirect to localized route
 - UrlRewriter: fixed greedy matching (`/en/` should not match `/men/`)
 - Fixed RuleSets could not be added or removed from a shipping method.
+- Fixed wrong SKU in order XML export if the order contains multiple variants of the same product.
 
 ## Smartstore 5.0.1